Refactor event handlers to separate modules
