+++ index.js
@@ -1,21 +1,12 @@
 import './modules/single-page-app-navigation.js';
 import BooksList from './modules/BookList.js';
 import renderAllBooks from './modules/render-books.js';
-import setupAddBookFormEventHandlers from './modules/set-up-add-book-form.js';
+import initializeAddBookFormEventHandlers from './modules/initialize-add-book-form-event-handlers.js';
+import initializeBookListEventHandlers from './modules/initialize-book-list-event-handlers.js';
 
 const awesomeBooks = new BooksList();
 
-setupAddBookFormEventHandlers(awesomeBooks);
-
-const booksListContainer = document.querySelector('#books');
-
-booksListContainer.addEventListener('click', (e) => {
-  if (e.target.classList.contains('remove-btn')) {
-    const id = e.target.dataset.bookId;
-    awesomeBooks.removeBook(id);
-    booksListContainer.innerHTML = '';
-    renderAllBooks(awesomeBooks.books);
-  }
-});
+initializeBookListEventHandlers(awesomeBooks);
+initializeAddBookFormEventHandlers(awesomeBooks);
 
 window.addEventListener('load', renderAllBooks(awesomeBooks.books));
