+++ modules/initialize-add-book-form-event-handlers.js
@@ -2,16 +2,16 @@ import renderAllBooks from './render-books.js';
 
 const form = document.getElementById('form');
 
-const setupAddBookFormEventHandlers = (booksList) => {
+const initializeAddBookFormEventHandlers = (booksList) => {
   form.addEventListener('submit', (e) => {
     e.preventDefault();
     const inputs = e.target.elements;
     const title = inputs.title.value;
     const author = inputs.author.value;
     booksList.addBook(title, author);
-    renderAllBooks();
+    renderAllBooks(booksList.books);
     e.target.reset();
   });
 };
 
-export default setupAddBookFormEventHandlers;
+export default initializeAddBookFormEventHandlers;
